added some exciting new bugs
I have the bare bones of block movement down.  Right now, the block
can be manipulated using the arrow keys, and can still be rotated
with the R key.

The problem: I need to streamline the drawing and redrawing process,
because I find myself needing to write the same logic into two or
three different methods which is... just terrible.  The most glaring
bug at the moment are two:
1. When a block is up against the wall and I hit R to rotate it, it
   doesn't rotate, but I can also no longer move it.  When I hit R
   again, it goes back to normal.

2. Blocks can still partially pass through the left wall.  I'm sure
   these are easy fixes, but I'm pressed for time and it'll be
   much easier when I can relax and fix this.