Fixed serious gameplay issues.
I hadn't mentioned it in the previous commit, but I had introduced
a bug that caused problems when a line was cleared.  Basically,
the active block wasn't being deactivated soon enough, which
caused it to be redrawn onscreen in a way it shouldn't have been.
Even parts of the block that should have been cleared were pushed
up and redrawn.

The cause of this bug was the order I was performing certain actions.
Previously, when the tick function was clearing out completed rows
after the active block landed, it hadn't been deactivated yet.
Since it was still active, it got redrawn for one more tick before
the active block was changed.